Fix missing GitHub stats when slug and key differ
Use a static site generator project's key instead of its slug to look up
project data.

The key is derived from the Markdown filename for the project. The slug
is derived from the project's title defined in its Markdown file. For
example the project defined in `content/projects/amsf.md` has the title
set to `Almace Scaffolding` in the Markdown file. The key for this
project is `amsf` whereas the slug for this project is
`almace-scaffolding`.

The project data is saved in `projectData` by the project's key. But
prior to this fix, the project's slug is used to look up the data. This
leads to missing GitHub stats for this project and other projects which
have their key different from their slug.

This fix involves looking up a project's data in `projectData` by key
instead of slug.