renamed metabake to intu.dev
Old: MetaBake was mostly a CLI tool, with a bit of WebAdmin.
The new version: intuition.dev still has CLI way to generate code, but is now primarily a WebAdmin tool.

With the new name, there is a new website/domain that matches the new name (old domain redirects to the new domain).